LPS-179756 Extract CPInstanceDetailsScreenNavigationCategory methods to a CPInstanceDetailsScreenNavigationEntry component and extends CPInstanceDetailsScreenNavigationCategory in it
